conntrack: remove filter_dump
filter_dump is specific to libnfnetlink but I will not use it. This filter is
done by putting CTA_MARK and CTA_MARK_MASK attr in request nlmsghdr. see
examples/netfilter/nfct-daemon.c in libmnl.

adversaria: I tried to filter dump by attaching bpf to socket. Bpf, sk_filter()
is actually called in netlink_dump() but the passed skb is not single. It is
filled multiple by ctnetlink_dump_table() so that it has no mean to
attach (normal) filter to dumping socket.